Add Builder provider interface
